requirements.txt
openai
google-generativeai
google-genai
python-dotenv
gemini
pandas
PyPDF2
faiss